packages/backend/requirements.txt
pillow-heif
Pillow
pytesseract
requests
resend
supabase
python-dotenv
